Fix CircleCI config for smoke tests
