+++ scripts/K6/support.js
@@ -1,0 +1,67 @@
+import http from "k6/http";
+import encoding from 'k6/encoding';
+
+export function importTest() {
+  return 1
+}
+
+function read_or_decode(value, suffix) {
+  if (value.includes(suffix)) {
+    return open(value);
+  } else {
+    return encoding.b64decode(value, 'std', 's');
+  }
+}
+
+export function read_certificate(profile) {
+  let cert_val = ""
+  let key_val = ""
+  let api_key_val = ""
+  switch (profile) {
+    case "MAIN":
+      cert_val = __ENV.FULL_ACCESS_CERT;
+      key_val = __ENV.FULL_ACCESS_KEY;
+      api_key_val = __ENV.FULL_ACCESS_API_KEY;
+      break
+    case "PROD":
+      cert_val = __ENV.SMOKE_TEST_CERT;
+      key_val = __ENV.SMOKE_TEST_KEY;
+      api_key_val = __ENV.SMOKE_TEST_API_KEY;
+      break
+    case "LAO":
+      cert_val = __ENV.LIMITED_ACCESS_CERT;
+      key_val = __ENV.LIMITED_ACCESS_KEY;
+      api_key_val = __ENV.LIMITED_ACCESS_API_KEY;
+      break
+    case "NOPERMS":
+      cert_val = __ENV.NO_ACCESS_CERT;
+      key_val = __ENV.NO_ACCESS_KEY;
+      api_key_val = __ENV.NO_ACCESS_API_KEY;
+      break
+    case "NOCERT":
+      cert_val = "";
+      key_val = "";
+      api_key_val = __ENV.NO_ACCESS_API_KEY;
+      break
+    default:
+      console.log("Unknown profile: " + profile);
+  }
+
+  return [
+    read_or_decode(cert_val, ".pem"),
+    read_or_decode(key_val, ".key"),
+    api_key_val
+  ]
+}
+
+/**
+ * Make a GET request to the API and validate that the http response code indicates syccess.
+ * @returns the http response object
+ */
+export function validate_get_request(path) {
+  const res = http.get(`${baseUrl}${path}`, httpParams);
+  check(res, {
+    [`GET ${path} successful`]: (r) => r.status < 400,
+  });
+  return res;
+}
